clean up / move code around
