--- web/app/cad/counters.js
@@ -1,7 +1,0 @@
-
-const Counters = {
-  solid : 0,
-  shared : 0
-};
-
-export default Counters;
